requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
delta-spark==3.2.0
filelock==3.16.1
frozenlist==1.5.0
gdown==5.2.0
idna==3.10
importlib_metadata==8.5.0
multidict==6.1.0
numpy==1.26.4
pandas==2.2.3
propcache==0.2.0
py4j==0.10.9.7
pyarrow==18.0.0
PySocks==1.7.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
ruff==0.7.2
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
structlog==24.4.0
tqdm==4.67.0
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
zipp==3.20.2
